backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
requests
pandas
openpyxl
python-multipart
python-jose[cryptography]
passlib[bcrypt]
supabase
python-dotenv
geojson
shapely
msal
azure-storage-blob
geoalchemy2
